+++ src/Knp/Bundle/KnpBundlesBundle/Command/KbAddBundleCommand.php
@@ -32,13 +32,14 @@ class KbAddBundleCommand extends ContainerAwareCommand
     {
         $container = $this->getContainer();
 
+        /* @var $updater Updater */
         $updater = $container->get('knp_bundles.updater');
-        
+
         if (!$input->getOption('no-publish')) {
             // manually set RabbitMQ producer
             $updater->setBundleUpdateProducer($container->get('old_sound_rabbit_mq.update_bundle_producer'));
         }
-        
+
         $updater->setOutput($output);
         $updater->setUp();