+++ src/Knp/Bundle/KnpBundlesBundle/Command/KbAddSearchedBundlesCommand.php
@@ -31,6 +31,7 @@ class KbAddSearchedBundlesCommand extends ContainerAwareCommand
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
+        /* @var $updater Updater */
         $updater = $this->getContainer()->get('knp_bundles.updater');
         $updater->setOutput($output);
         $updater->setUp();